Update
Bringing the latest updates for improving the use of commands for both
pokemon and waifu.

Note that this versino fo Pokemon will not work with old saves, it will
generate a new one.

I also added a fix for the null "2" error, please contact me if this
error is still happening.